Replace ProcessRouter with AbstractProcessDispatcher
Removed ProcessRouter and introduced AbstractProcessDispatcher to refactor process handling logic. The new abstract class centralizes process management and scheduling, simplifying the interface and removing path-based handler registration.
